refactor: move document constants to constants folder
