+++ src/core/constants/documentType.constant.tsx
